fix(arch): UX reload→callback, alert→state, auth split, isWednesdayParis, Stripe idempotence
